+++ api/src/app.module.ts
@@ -3,11 +3,12 @@ import { APP_FILTER } from "@nestjs/core";
 import { PrismaModule } from "./prisma/prisma.module";
 import { AccountsModule } from "./accounts/accounts.module";
 import { CategoriesModule } from "./categories/categories.module";
+import { TransactionsModule } from "./transactions/transactions.module";
 import { HealthController } from "./health/health.controller";
 import { HttpExceptionFilter } from "./common/filters/http-exception.filter";
 
 @Module({
-  imports: [PrismaModule, AccountsModule, CategoriesModule],
+  imports: [PrismaModule, AccountsModule, CategoriesModule, TransactionsModule],
   controllers: [HealthController],
   providers: [
     {